refactor: :truck: Move most of GPU code into one level higher
`git mv` not behaving well on moving `mod.rs` so will move this in
separate commit.
