+++ src/gpu/main_windowed.rs
@@ -1,5 +1,5 @@
 #[allow(unused_imports)]
-use pdb_tui::gpu::pdb_gpu::run_windowed::run;
+use pdb_tui::gpu::run_windowed::run;
 
 fn main() {
     pollster::block_on(run());
