+++ src/gpu/state_windowless.rs
@@ -2,8 +2,8 @@ use image::{ImageBuffer, Rgba};
 use std::iter;
 use winit::dpi::PhysicalSize;
 
-use crate::gpu::pdb_gpu::model::{DrawLight, DrawModel};
-use crate::gpu::pdb_gpu::{InnerState, State};
+use crate::gpu::model::{DrawLight, DrawModel};
+use crate::gpu::{InnerState, State};
 
 const FONT_ASPECT_RATIO: f32 = 2.0;
