+++ src/gpu/trivial_compute.wgsl
